+++ tests/accelerators/test_accelerator_connector.py
@@ -533,7 +533,7 @@ def test_accelerator_cpu_with_devices(devices, plugin):
 
     trainer = Trainer(accelerator="cpu", devices=devices)
 
-    assert trainer.num_processes == devices
+    assert trainer.num_devices == devices
     assert isinstance(trainer.strategy, plugin)
     assert isinstance(trainer.accelerator, CPUAccelerator)
 
@@ -545,7 +545,7 @@ def test_accelerator_cpu_with_num_processes_priority():
     with pytest.warns(UserWarning, match="The flag `devices=8` will be ignored,"):
         trainer = Trainer(accelerator="cpu", devices=8, num_processes=num_processes)
 
-    assert trainer.num_processes == num_processes
+    assert trainer.num_devices == num_processes
 
 
 @RunIf(min_gpus=2)
@@ -583,7 +583,7 @@ def test_validate_accelerator_and_devices():
 
     trainer = Trainer(accelerator="ddp_cpu", devices=2)
     assert isinstance(trainer.accelerator, CPUAccelerator)
-    assert trainer.num_processes == 2
+    assert trainer.num_devices == 2
 
 
 def test_set_devices_if_none_cpu():
@@ -962,7 +962,6 @@ def test_unsupported_ipu_choice(mock_ipu_acc_avail, monkeypatch):
 def test_devices_auto_choice_cpu(is_ipu_available_mock, is_tpu_available_mock, is_gpu_available_mock):
     trainer = Trainer(accelerator="auto", devices="auto")
     assert trainer.num_devices == 1
-    assert trainer.num_processes == 1
 
 
 @mock.patch("torch.cuda.is_available", return_value=True)